move protocol creation methods into create_protocols.py
